Move filename logic to output parameters (fix "tell, don't ask" violation)
