+++ src/main/java/com/vnaskos/pdfguru/execution/OutputParameters.java
@@ -1,28 +1,47 @@
 package com.vnaskos.pdfguru.execution;
 
+import java.io.File;
+
 /**
  *
  * @author Vasilis Naskos
  */
 public class OutputParameters {
 
+    private static final int INITIAL_INDEX = 1;
     static final float COMPRESSED_MAX = 0.0f;
     static final float UNCOMPRESSED_MAX = 1.0f;
 
-    private final String outputFile;
+    private final String outputFilePath;
+    private int index = INITIAL_INDEX;
 
     private float compression = 0.5f;
     private boolean singleFileOutput = true;
 
-    public OutputParameters(String outputFile) {
-        this.outputFile = outputFile;
+    public OutputParameters(String outputFilePath) {
+        this.outputFilePath = outputFilePath;
     }
 
-    public String getOutputFile() {
-        return outputFile;
+    String getUniqueOutputFileName() {
+        String tmpOutputFilePath = outputFilePath;
+
+        if(outputFilePath.toLowerCase().endsWith(".pdf")) {
+            tmpOutputFilePath = outputFilePath.substring(0, outputFilePath.length()-4);
+        }
+
+        String uniqueOutputFilePath;
+        do {
+            uniqueOutputFilePath = String.format("%s_%d.pdf", tmpOutputFilePath, index++);
+        } while (fileExists(uniqueOutputFilePath));
+
+        return uniqueOutputFilePath;
     }
 
-    public float getCompression() {
+    boolean fileExists(String testOut) {
+        return (new File(testOut)).exists();
+    }
+
+    float getCompression() {
         return compression;
     }
 
@@ -40,7 +59,7 @@ public class OutputParameters {
         this.singleFileOutput = false;
     }
 
-    public boolean isSingleFileOutput() {
+    boolean isSingleFileOutput() {
         return singleFileOutput;
     }
 }
