+++ src/test/java/com/vnaskos/pdfguru/execution/OutputParametersTest.java
@@ -3,6 +3,8 @@ package com.vnaskos.pdfguru.execution;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.spy;
 
 public class OutputParametersTest {
 
@@ -36,5 +38,35 @@ public class OutputParametersTest {
         parameters.setCompression(aValueAboveOne);
 
         assertThat(parameters.getCompression()).isEqualTo(OutputParameters.UNCOMPRESSED_MAX);
+    }
+
+    @Test
+    public void createOutputNameCombiningOutputPathIndexAndExtension() {
+        OutputParameters outputParameters = new OutputParameters(AN_OUTPUT_FILEPATH);
+
+        String outputFileName = outputParameters.getUniqueOutputFileName();
+
+        assertThat(outputFileName).isEqualTo("/somewhere/over/the/rainbow_1.pdf");
+    }
+
+    @Test
+    public void createOutputNameWhenOutputPathContainsExtensionMaintainExtensionLast() {
+        OutputParameters outputParameters = new OutputParameters(AN_OUTPUT_FILEPATH + ".pdf");
+
+        String outputFileName = outputParameters.getUniqueOutputFileName();
+
+        assertThat(outputFileName).isEqualTo("/somewhere/over/the/rainbow_1.pdf");
+    }
+
+    @Test
+    public void createUniqueOutputNameWhenFilesWithTheFirstTwoIndicesAlreadyExists() {
+        OutputParameters outputParameters = spy(new OutputParameters(AN_OUTPUT_FILEPATH));
+
+        doReturn(true).when(outputParameters).fileExists("/somewhere/over/the/rainbow_1.pdf");
+        doReturn(true).when(outputParameters).fileExists("/somewhere/over/the/rainbow_2.pdf");
+
+        String outputFileName = outputParameters.getUniqueOutputFileName();
+
+        assertThat(outputFileName).isEqualTo("/somewhere/over/the/rainbow_3.pdf");
     }
 }
